+++ doc/changelog.rst
@@ -6,6 +6,11 @@ Changelog
 =========
 2.8.0
 -----
+API change
+~~~~~~~~~~
+- ``Minuit.mncontour`` now raises ``RuntimeError`` instead of ``ValueError`` if it is
+  not called at a valid minimum point
+
 New features
 ~~~~~~~~~~~~
 - ``Minuit.mncontour`` can now be called at any point without running a minimiser before,